Add a PAM module
This is necessary unfortunately due to shortcomings in the
fprintd module.

pam_fprintd.so always just blindly uses fingerprint authentication.
It does not account for use cases where fingerprint authentication
is not recommended.

For instance, when a user logs in with an encrypted home
directory, it is necessary in this case to perform a password
authentication, otherwise auth succeeds but home dir decryption
fails.

pam_fingwit.so just detects these cases and either returns
PAM_IGNORE when it's safe for pam_fprintd.so to perform a fingerprint
authentication, or PAM_AUTHINFO_UNAVAIL when pam_fprint.so should be
skipped and pam_unix.so used instead.